# add isequal and hash methods to check tuple in BoundaryIndex Sets without type piracy 