--- a/src/Grid/grid.jl
+++ b/src/Grid/grid.jl
@@ -765,11 +765,11 @@ for INDEX in (:VertexIndex, :EdgeIndex, :FaceIndex)
         #To be able to do a,b = faceidx
         Base.iterate(I::($INDEX), state::Int=1) = (state==3) ?  nothing : (I[state], state+1)
 
-        #Necessary to check if, e.g. `(cellid, faceidx) in faceset`
-        Base.isequal(x::$INDEX, y::$INDEX) = x.idx[1] == y.idx[1] && x.idx[2] == y.idx[2]
+        # Necessary to check if, e.g. `(cellid, faceidx) in faceset`
+        Base.isequal(x::$INDEX, y::$INDEX) = x.idx == y.idx
         Base.isequal(x::Tuple{Int, Int}, y::$INDEX) = x[1] == y.idx[1] && x[2] == y.idx[2]
         Base.isequal(y::$INDEX, x::Tuple{Int, Int}) = x[1] == y.idx[1] && x[2] == y.idx[2]
-        Base.hash(x::$INDEX,h::UInt) = Base.hash(x.idx,h)
+        Base.hash(x::$INDEX, h::UInt) = hash(x.idx, h)
     end
 end