@koehlerson 2023-11-03
> Nice!
> 
> Are there any performance differences?

master
```
julia> @benchmark in((1,2),testset)
BenchmarkTools.Trial: 10000 samples with 989 evaluations.
 Range (min … max):  48.218 ns … 69.210 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     48.507 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   48.717 ns ±  0.988 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆▆█▇█▆▄                              ▁ ▁▁                   ▂
  ███████▆▃▃▁▄▄▅▃▅▄▅▄▃▁▅▃▃▄▃▁▄▄▅▄▄▁▅▆▇█████▇▇█▇██▆▇▇▇▆▇▇▆▆▆▆▆ █
  48.2 ns      Histogram: log(frequency) by time        53 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

PR
```
julia> @benchmark in((1,2),testset)
BenchmarkTools.Trial: 10000 samples with 994 evaluations.
 Range (min … max):  31.995 ns … 63.816 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     32.391 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   32.565 ns ±  1.520 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃  █▃                                                       ▁
  █▅▃██▆▃▃▁▁▁▃▁▁▁▁▁▁▃▁▁▁▁▁▁▁▃▁▁▁▁▁▃▁▁▁▁▁▃▄▃▃▆▅▄▆▅▄▄▅▅▅▅▃▄▇▄▃▄ █
  32 ns        Histogram: log(frequency) by time      38.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
 
```