@koehlerson 2023-11-06
I just noticed the following, which I encountered in the p4est branch as well. If we hash now `FaceIndex(1,2)` and `EdgeIndex(1,2)` the hash will be the same. This can be problematic if you have `Set{Union{FaceIndex,EdgeIndex}}` and do some set operations (which I think rely on the hashes and not only the `isequal` operation). In the p4est branch I just wrapped the `Tuple{Int,Int}` in a `Tuple{Int,Tuple{Int,Int}}` where the first int is sort of a category integer, but not sure if relevant