@KnutAM 2023-11-06
Good point - did not think of this! 

I tried to understand what's going on, and AFAIU it works because `isequal(FaceIndex(1,2), EdgeIndex(1,2))` is `false`, such that the following check give `false` and the loop continues until eventually returning a negative index (indicating that the key wasn't found), for a set without `FaceIndex(1,2)` but with `EdgeIndex(1,2)`. 

https://github.com/JuliaLang/julia/blob/6d4f40957c6e68ecfd1dc13115beadff493406fd/base/dict.jl#L289

That being said, I think we need @fredrikekre or @KristofferC to confirm that this way is ok. The [rule for defining the hash](https://docs.julialang.org/en/v1/base/base/#Base.hash): `"Compute an integer hash code such that isequal(x,y) implies hash(x)==hash(y)"`, use *implies*, so to my understanding the implementation in this PR follows the docstring.