@koehlerson 2023-11-06
yeah I'm just unsure if the statement needs to hold the other way around too. So should `hash(x) == hash(y)` imply `isequal(x,y)`. If so, then we should change it.