@KnutAM 2023-11-06
> yeah I'm just unsure if the statement needs to hold the other way around too. So should hash(x) == hash(y) imply isequal(x,y). If so, then we should change it.

At least the [docstring](https://docs.julialang.org/en/v1/base/base/#Base.isequal) for `isequal` is consistent with our interpretation: `"isequal(x,y) must imply that hash(x) == hash(y)."`